Rename main and entity dashboard terminology.
