+++ assets/sass/admin.scss
@@ -133,7 +133,7 @@
 @import "components/dashboard/googlesitekit-signin-box";
 @import "components/dashboard/googlesitekit-dashboard-charts";
 @import "components/dashboard/googlesitekit-dashboard-module-header";
-@import "components/dashboard/googlesitekit-dashboard-single-url";
+@import "components/dashboard/googlesitekit-entity-dashboard";
 @import "components/dashboard/googlesitekit-DashboardPageSpeed";
 @import "components/dashboard/googlesitekit-ReportMetric";
