+++ assets/sass/components/dashboard/_googlesitekit-entity-dashboard.scss
@@ -16,13 +16,13 @@
  * limitations under the License.
  */
 
-.googlesitekit-dashboard-single-url {
+.googlesitekit-entity-dashboard {
 
-	.googlesitekit-dashboard-single-url__heading {
+	.googlesitekit-entity-dashboard__heading {
 		margin: 10px 0 0;
 	}
 
-	.googlesitekit-dashboard-single-url__entity-header {
+	.googlesitekit-entity-dashboard__entity-header {
 		// Use the same spacing as is used in layout grids.
 		@each $size in map-keys($mdc-layout-grid-columns) {
 
@@ -34,7 +34,7 @@
 		}
 	}
 
-	.googlesitekit-dashboard-single-url__title {
+	.googlesitekit-entity-dashboard__title {
 		margin: 0 0 10px;
 	}
 }
