+++ tests/e2e/specs/modules/search-console/fixtures/entity-dashboard.js
