+++ tests/e2e/specs/modules/search-console/fixtures/main-dashboard.js
